+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -25,6 +25,7 @@ import {
 import { Log } from '../../../log';
 import { FileNotifier } from '../../../utils/FileNotifier';
 import { env } from '../../../utils/env';
+import { CommandError } from '../../../utils/errors';
 import { installExitHooks } from '../../../utils/exit';
 import { isInteractive } from '../../../utils/interactive';
 import { loadTsConfigPathsAsync, TsConfigPaths } from '../../../utils/tsconfig/loadTsConfigPaths';
@@ -67,7 +68,7 @@ function withWebPolyfills(
           return `global.$$require_external = typeof window === "undefined" ? require : () => null;`;
         } else {
           // Wrap in try/catch to support Android.
-          return 'try { global.$$require_external = typeof expo === "undefined" ? eval("require") : (moduleId) => { throw new Error(`Node.js standard library module ${moduleId} is not available in this JavaScript environment`);} } catch { global.$$require_external = (moduleId) => { throw new Error(`Node.js standard library module ${moduleId} is not available in this JavaScript environment`);} }';
+          return 'try { global.$$require_external = typeof expo === "undefined" ? require : (moduleId) => { throw new Error(`Node.js standard library module ${moduleId} is not available in this JavaScript environment`);} } catch { global.$$require_external = (moduleId) => { throw new Error(`Node.js standard library module ${moduleId} is not available in this JavaScript environment`);} }';
         }
       })()
     );
@@ -148,12 +149,36 @@ export function withExtendedResolver(
     Log.warn(`Experimental React Canary version is enabled.`);
   }
 
-  // Get the `transformer.assetRegistryPath`
-  // this needs to be unified since you can't dynamically
-  // swap out the transformer based on platform.
-  const assetRegistryPath = fs.realpathSync(
-    path.resolve(resolveFrom(config.projectRoot, '@react-native/assets-registry/registry.js'))
-  );
+  let _assetRegistryPath: string | null = null;
+
+  // Fetch this lazily for testing purposes.
+  function getAssetRegistryPath() {
+    if (_assetRegistryPath) {
+      return _assetRegistryPath;
+    }
+
+    // Get the `transformer.assetRegistryPath`
+    // this needs to be unified since you can't dynamically
+    // swap out the transformer based on platform.
+    if (
+      config.transformer.assetRegistryPath &&
+      path.isAbsolute(config.transformer.assetRegistryPath)
+    ) {
+      _assetRegistryPath = fs.realpathSync(config.transformer.assetRegistryPath);
+      return _assetRegistryPath;
+    }
+
+    const assetRegistryPath = fs.realpathSync(
+      path.resolve(
+        resolveFrom(
+          config.projectRoot,
+          config.transformer.assetRegistryPath ?? '@react-native/assets-registry/registry.js'
+        )
+      )
+    );
+    _assetRegistryPath = assetRegistryPath;
+    return assetRegistryPath;
+  }
 
   const defaultResolver = metroResolver.resolve;
   const resolver = isFastResolverEnabled
@@ -173,12 +198,22 @@ export function withExtendedResolver(
     },
   };
 
-  const universalAliases: [RegExp, string][] = [];
+  let _universalAliases: [RegExp, string][] | null;
 
-  // This package is currently always installed as it is included in the `expo` package.
-  if (resolveFrom.silent(config.projectRoot, '@expo/vector-icons')) {
-    debug('Enabling alias: react-native-vector-icons -> @expo/vector-icons');
-    universalAliases.push([/^react-native-vector-icons(\/.*)?/, '@expo/vector-icons$1']);
+  function getUniversalAliases() {
+    if (_universalAliases) {
+      return _universalAliases;
+    }
+
+    _universalAliases = [];
+
+    // This package is currently always installed as it is included in the `expo` package.
+    if (resolveFrom.silent(config.projectRoot, '@expo/vector-icons')) {
+      debug('Enabling alias: react-native-vector-icons -> @expo/vector-icons');
+      _universalAliases.push([/^react-native-vector-icons(\/.*)?/, '@expo/vector-icons$1']);
+    }
+
+    return _universalAliases;
   }
 
   const preferredMainFields: { [key: string]: string[] } = {
@@ -262,6 +297,33 @@ export function withExtendedResolver(
     };
   }
 
+  // If Node.js pass-through, then remap to a module like `module.exports = $$require_external(<module>)`.
+  // If module should be shimmed, remap to an empty module.
+  const externals: {
+    match: (context: ResolutionContext, moduleName: string, platform: string | null) => boolean;
+    replace: 'empty' | 'node';
+  }[] = [
+    {
+      match: (context: ResolutionContext, moduleName: string) => {
+        if (
+          // Disable internal externals when exporting for production.
+          context.customResolverOptions.exporting ||
+          // These externals are only for Node.js environments.
+          !isServerEnvironment(context.customResolverOptions?.environment)
+        ) {
+          return false;
+        }
+
+        // Extern these modules in standard Node.js environments in development to prevent API routes side-effects
+        // from leaking into the dev server process.
+        return /^(source-map-support(\/.*)?|react|react-native-helmet-async|@radix-ui\/.+|@babel\/runtime\/.+|react-dom(\/.+)?|debug|acorn-loose|acorn|css-in-js-utils\/lib\/.+|hyphenate-style-name|color|color-string|color-convert|color-name|fontfaceobserver|fast-deep-equal|query-string|escape-string-regexp|invariant|postcss-value-parser|memoize-one|nullthrows|strict-uri-encode|decode-uri-component|split-on-first|filter-obj|warn-once|simple-swizzle|is-arrayish|inline-style-prefixer\/.+)$/.test(
+          moduleName
+        );
+      },
+      replace: 'node',
+    },
+  ];
+
   const metroConfigWithCustomResolver = withMetroResolvers(config, [
     // Mock out production react imports in development.
     (context: ResolutionContext, moduleName: string, platform: string | null) => {
@@ -311,11 +373,6 @@ export function withExtendedResolver(
         context.customResolverOptions?.environment === 'node' ||
         context.customResolverOptions?.environment === 'react-server';
 
-      if (platform !== 'web' && !isServer) {
-        // This is a web/server-only feature, we may extend the shimming to native platforms in the future.
-        return null;
-      }
-
       const moduleId = isNodeExternal(moduleName);
       if (!moduleId) {
         return null;
@@ -329,6 +386,12 @@ export function withExtendedResolver(
         // Perform optional resolve first. If the module doesn't exist (no module in the node_modules)
         // then we can mock the file to use an empty module.
         const result = getOptionalResolver(context, platform)(moduleName);
+
+        if (!result && platform !== 'web') {
+          // Preserve previous behavior where native throws an error on node.js internals.
+          return null;
+        }
+
         return (
           result ?? {
             // In this case, mock the file to use an empty module.
@@ -350,6 +413,42 @@ export function withExtendedResolver(
       };
     },
 
+    // Custom externals support
+    (context: ResolutionContext, moduleName: string, platform: string | null) => {
+      // We don't support this in the resolver at the moment.
+      if (moduleName.endsWith('/package.json')) {
+        return null;
+      }
+
+      for (const external of externals) {
+        if (external.match(context, moduleName, platform)) {
+          if (external.replace === 'empty') {
+            debug(`Redirecting external "${moduleName}" to "${external.replace}"`);
+            return {
+              type: external.replace,
+            };
+          } else if (external.replace === 'node') {
+            const contents = `module.exports=$$require_external('${moduleName}')`;
+            const virtualModuleId = `\0node:${moduleName}`;
+            debug('Virtualizing Node.js (custom):', moduleName, '->', virtualModuleId);
+            getMetroBundlerWithVirtualModules(getMetroBundler()).setVirtualModule(
+              virtualModuleId,
+              contents
+            );
+            return {
+              type: 'sourceFile',
+              filePath: virtualModuleId,
+            };
+          } else {
+            throw new CommandError(
+              `Invalid external alias type: "${external.replace}" for module "${moduleName}" (platform: ${platform}, originModulePath: ${context.originModulePath})`
+            );
+          }
+        }
+      }
+      return null;
+    },
+
     // Basic moduleId aliases
     (context: ResolutionContext, moduleName: string, platform: string | null) => {
       // Conditionally remap `react-native` to `react-native-web` on web in
@@ -359,7 +458,7 @@ export function withExtendedResolver(
         return getStrictResolver(context, platform)(redirectedModuleName);
       }
 
-      for (const [matcher, alias] of universalAliases) {
+      for (const [matcher, alias] of getUniversalAliases()) {
         const match = moduleName.match(matcher);
         if (match) {
           const aliasedModule = alias.replace(
@@ -380,6 +479,17 @@ export function withExtendedResolver(
     (context: ResolutionContext, moduleName: string, platform: string | null) => {
       const doResolve = getStrictResolver(context, platform);
 
+      if (
+        platform === 'web' &&
+        context.originModulePath.match(/node_modules[\\/]react-native-web[\\/]/) &&
+        moduleName.includes('/modules/AssetRegistry')
+      ) {
+        return {
+          type: 'sourceFile',
+          filePath: getAssetRegistryPath(),
+        };
+      }
+
       const result = doResolve(moduleName);
 
       if (result.type !== 'sourceFile') {
@@ -387,15 +497,7 @@ export function withExtendedResolver(
       }
 
       if (platform === 'web') {
-        // Replace the web resolver with the original one.
-        // This is basically an alias for web-only.
-        // TODO: Drop this in favor of the standalone asset registry module.
-        if (shouldAliasAssetRegistryForWeb(platform, result)) {
-          // @ts-expect-error: `readonly` for some reason.
-          result.filePath = assetRegistryPath;
-        }
-
-        if (platform === 'web' && result.filePath.includes('node_modules')) {
+        if (result.filePath.includes('node_modules')) {
           // Replace with static shims
 
           const normalName = normalizeSlashes(result.filePath)
@@ -462,9 +564,15 @@ export function withExtendedResolver(
 
         context.unstable_enablePackageExports = true;
         context.unstable_conditionsByPlatform = {};
-        // Node.js runtimes should only be importing main at the moment.
-        // This is a temporary fix until we can support the package.json exports.
-        context.mainFields = ['main', 'module'];
+
+        if (platform === 'web') {
+          // Node.js runtimes should only be importing main at the moment.
+          // This is a temporary fix until we can support the package.json exports.
+          context.mainFields = ['main', 'module'];
+        } else {
+          // In Node.js + native, use the standard main fields.
+          context.mainFields = ['react-native', 'main', 'module'];
+        }
 
         // Enable react-server import conditions.
         if (context.customResolverOptions?.environment === 'react-server') {
@@ -487,20 +595,6 @@ export function withExtendedResolver(
   return withMetroErrorReportingResolver(metroConfigWithCustomContext);
 }
 
-/** @returns `true` if the incoming resolution should be swapped on web. */
-export function shouldAliasAssetRegistryForWeb(
-  platform: string | null,
-  result: Resolution
-): boolean {
-  return (
-    platform === 'web' &&
-    result?.type === 'sourceFile' &&
-    typeof result?.filePath === 'string' &&
-    normalizeSlashes(result.filePath).endsWith(
-      'react-native-web/dist/modules/AssetRegistry/index.js'
-    )
-  );
-}
 /** @returns `true` if the incoming resolution should be swapped. */
 export function shouldAliasModule(
   input: {